moving getManagementCanister into Actor to avoid circular deps
