+++ packages/agent/src/index.ts
@@ -7,7 +7,6 @@ export * from './certificate';
 export * from './agent/http/transforms';
 export * from './agent/http/types';
 export * from './canisters/asset';
-export * from './canisters/management';
 export * from './fetch_candid';
 export * from './request_id';
 export * from './utils/bls';
